fix url api list comments
