+++ yashoes/product/urls.py
@@ -5,6 +5,7 @@ from yashoes.product import views
 urlpatterns = [
     path('', views.ProductsAPIView.as_view()),
     path('<int:pk>', views.ProductDetail.as_view()),
+    path('<int:product_id>/comments', views.GetCommentView.as_view())
 ]
 
 # urlpatterns = format_suffix_patterns(urlpatterns)
