+++ yashoes/urls.py
@@ -4,6 +4,5 @@ urlpatterns = [
     path('', include('yashoes.auth.urls')),
     path('products/', include('yashoes.product.urls')),
     path('users/', include('yashoes.user.urls')),
-    path('comments/', include('yashoes.comment.urls')),
     path('', include('yashoes.transaction.urls')),
 ]
